节点宕机后，重启。或失去联系zk后又重新连接了。都视为新节点加入集群。加入前需要清空所有记录，避免有重复数据在集群中

+++ easyTask-L-core/src/main/java/com/github/liuche51/easyTask/dao/ScheduleBakDao.java
@@ -94,6 +94,7 @@ public class ScheduleBakDao {
         SqliteHelper.executeUpdateForSync(sql);
     }
     public static void deleteBySources(String[] sources) throws SQLException, ClassNotFoundException {
+        if(sources==null||sources.length==0) return;
        String conditionStr=SqliteHelper.getInConditionStr(sources);
         String sql = "delete FROM schedule_bak where source not in" + conditionStr + ";";
         SqliteHelper.executeUpdateForSync(sql);
